Break up utils.hpp for better organization and faster compilation
