+++ dwave/optimization/src/nodes/binaryop.cpp
@@ -22,7 +22,6 @@
 #include <vector>
 
 #include "_state.hpp"
-#include "dwave-optimization/utils.hpp"
 
 namespace dwave::optimization {
