+++ dwave/optimization/src/nodes/naryop.cpp
@@ -21,7 +21,6 @@
 #include <type_traits>
 
 #include "_state.hpp"
-#include "dwave-optimization/utils.hpp"
 
 namespace dwave::optimization {
