+++ meson.build
@@ -47,9 +47,9 @@ dwave_optimization_src = [
     'dwave/optimization/src/nodes/testing.cpp',
     'dwave/optimization/src/nodes/unaryop.cpp',
     'dwave/optimization/src/array.cpp',
+    'dwave/optimization/src/fraction.cpp',
     'dwave/optimization/src/graph.cpp',
     'dwave/optimization/src/simplex.cpp',
-    'dwave/optimization/src/utils.cpp',
 ]
 
 if host_machine.system() == 'windows'
