+++ tests/cpp/meson.build
@@ -28,12 +28,13 @@ tests_all = executable(
         'nodes/test_statistics.cpp',
         'nodes/test_unaryop.cpp',
         'test_array.cpp',
+        'test_double_kahan.cpp',
+        'test_fraction.cpp',
         'test_functional.cpp',
         'test_graph.cpp',
         'test_iterators.cpp',
         'test_simplex.cpp',
         'test_type_list.cpp',
-        'test_utils.cpp',
     ],
     dependencies: [
         catch2,
