+++ tests/cpp/test_double_kahan.cpp
@@ -1,4 +1,4 @@
-// Copyright 2024 D-Wave Systems Inc.
+// Copyright 2025 D-Wave
 //
 //    Licensed under the Apache License, Version 2.0 (the "License");
 //    you may not use this file except in compliance with the License.
@@ -14,14 +14,13 @@
 
 #include "catch2/catch_test_macros.hpp"
 #include "catch2/matchers/catch_matchers_all.hpp"
-#include "dwave-optimization/array.hpp"
-#include "dwave-optimization/utils.hpp"
+#include "double_kahan.hpp"
 
 using Catch::Matchers::RangeEquals;
 
 namespace dwave::optimization {
 
-TEST_CASE("CompensatedSum") {
+TEST_CASE("double_kahan") {
     GIVEN("An uninitialized struct for compensated sum") {
         double_kahan a;
 
@@ -85,50 +84,6 @@ TEST_CASE("CompensatedSum") {
     STATIC_REQUIRE(5 == double_kahan(5));
     STATIC_REQUIRE(6.0 != double_kahan(5));
     STATIC_REQUIRE(6 != double_kahan(5));
-}
-
-TEST_CASE("Test fraction") {
-    // everything can be done with constexpr so we can static require everything
-
-    STATIC_REQUIRE(fraction(3) == fraction(3, 1));
-    STATIC_REQUIRE(fraction(3) == fraction(6, 2));
-    STATIC_REQUIRE(fraction(3) == fraction(-12, -4));
-
-    STATIC_REQUIRE(fraction(0) == fraction(0, 5));
-    STATIC_REQUIRE(fraction(1, 5) * 0 == fraction(0));
-    STATIC_REQUIRE(fraction(1, 2) * 0 == fraction(0, 2) / 2);
-
-    STATIC_REQUIRE(fraction(3) == 3);
-    STATIC_REQUIRE(3 == fraction(3));
-    STATIC_REQUIRE(fraction(3, -1) == -3);
-    STATIC_REQUIRE(fraction(-3, 1) == -3);
-    STATIC_REQUIRE(fraction(-3, -1) == 3);
-
-    STATIC_REQUIRE(fraction(3) != 4);
-    STATIC_REQUIRE(4 != fraction(3));
-    STATIC_REQUIRE(fraction(3, -1) != 3);
-    STATIC_REQUIRE(fraction(-3, 1) != 3);
-    STATIC_REQUIRE(fraction(100, 50) == 2);
-    STATIC_REQUIRE(fraction(100, 50).numerator() == 2);
-    STATIC_REQUIRE(fraction(100, 50).denominator() == 1);
-    STATIC_REQUIRE(fraction(50, 100).numerator() == 1);
-    STATIC_REQUIRE(fraction(50, 100).denominator() == 2);
-    STATIC_REQUIRE(fraction(50, -100).numerator() == -1);
-    STATIC_REQUIRE(fraction(50, -100).denominator() == 2);
-
-    STATIC_REQUIRE(static_cast<double>(fraction(1, 2)) == .5);
-
-    STATIC_REQUIRE(fraction() == 0);
-
-    GIVEN("An uninitialized fraction") {
-        fraction a;
-
-        WHEN("We assign a value to it") {
-            a = 5;
-
-            THEN("It gets the value we expect") { CHECK(a == 5); }
-        }
-    }
 }
 
 }  // namespace dwave::optimization
